+++ backend/crm_api/src/helpers/mod.rs
@@ -1,0 +1,1 @@
+pub mod query_utils;
