+++ backend/crm_api/src/helpers/query_utils.rs
@@ -1,0 +1,7 @@
+pub fn process_optional_param(param: Option<Vec<String>>) -> Vec<String> {
+    let mut processed_param = vec![];
+    if let Some(populated_param) = param {
+        processed_param = populated_param
+    }
+    processed_param
+}
