+++ backend/crm_api/src/lib.rs
@@ -1,4 +1,5 @@
 mod routes;
+mod helpers;
 use axum::Router;
 use routes::create_routes;
 use sqlx::postgres::PgPoolOptions;
